[nxpfmac] Disconnect in connection destructor
The ClientConnection destructor would cancel any ongoing connection
attempts and make sure that any ongoing disconnects would complete. But
it would not actually perform a disconnect if it was connected.

Make sure we revert to a disconnected state if a client connection
object is destroyed. By waiting for the disconnect in progress state we
ensure that the disconnect either completed or wasn't necessary.

Test: Unit tests, acts start/stop client tests
Change-Id: I97d3c2eb041f74f15d963602903006271e77ad50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735866
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>