[build][fidl] Remove implicit _hlcpp allowlist
This CL removes the final two entries from _party from
_implicit_hlcpp_bindings_visibility, //third_party and //vendor. Now
the only way to get HLCPP bindings is to depend on the _hlcpp target.

This is blocked on third_party CLs landing:

* openweave-core: I1f4cc2389dc350a3a485b80fe2a9932bff1e80a7
* mesa: Ifac3f9ad3f74cedef77e447456d991ae798045bb
* Vulkan-Tools: Ic58248e491ae9ee8064baee2122bdeeab1a6222a

and also a couple vendor CLs.

Earlier steps in this transition:

* Remove //{examples,sdk,tools,zircon}
  Ie9d73b0a1ad78e7d34ee017017bacbaee34001d0
* Remove //src
  Ic30cc1914ada366a1cd3182c815d23311fcaba21

Fixed: 108354
Change-Id: Ie7a09d316b672a37d3bfa9f48ef81a1eb3715dd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734855
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Adam Barth <abarth@google.com>