[emu] Make --gpu and --device ffx_config-backed
Part of the problems we've been having with emulation on cloudtops can be
resolved through simple configuration, but it's not obvious to new users
how to change the config. We've talked about adding checks to a
"preflight" tool and setting the configuration according to the user's
host device capabilities.

This CL makes a few of the parameters to `ffx emu start` persistent
configuration options in ffx_config. This will allow the preflight tool
to determine the right config for the user's host, and set those values
as defaults in the user's ffx_config so they'll be propagated to the
`emu start` command automatically. They can still be overridden by the
user at any time on the CLI, or by modifying the ffx_config values.

Bug: 110080
Change-Id: I82227b890108c6a651de40451b77bc82282e5847
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736102
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>