[acpi] Fix racy metadata addition.
When we switched to using a passthrough device, we left metadata
initialisation in the same place. This was actually racy, because
sometimes the passthrough would get bound to before the parent's init
hook got called, which meant that the child driver saw no metadata and
failed to initialise.

We only observed impact from this on a workstation_userdebug debug build
on m3 atlases.

Test: boot workstation_userdebug on m3 atlas, try touchpad/mouse/login
Bug: 111140
Change-Id: I1ba021b323ee7743bc0bc00b16065a6c5cabca0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735436
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>