[fidl][scripts] Extend scaffolding create command
The previous change added the scaffolding script, and endowed it with
one command: "create", which (as of that chagne) only works to the end
of creating a *new* canonical example series' baseline case. This change
extends the create command to also allow for the creation of variants on
an already existing series. The `scaffold.py create` invocation must
take either a `--new` flag (indicating a new series), or a
`--from=SERIES` flag, indicating a variant on an existing series.

Test: manual, used to generate fxr/737545
Bug: 111614
Change-Id: Iea4eb83845040853da8d85f950934f6dfdd6e478
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737209
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>