[fidl][scripts] Add `scaffold document` command
This command takes an already created canonical example variant and adds
documentation to it. That doesn't mean it writes the actual
documentation (that would be quite a feat...), but it does create all of
the documentation templates (like devsite widgets, index entries, etc),
and leaves `TODO(DONOTSUBMIT)` comments for the user. The output of the
command then tells the user where all of these comments are located, so
that they may replace them with the appropriate text.

Test: manual, used to generate fxr/737546 (PS1)
Bug: 111614
Change-Id: I16342ec9c539d98f7860268a232a747ade8d762f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737210
Reviewed-by: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>