[fidl][scripts] Add `scaffold implement` for Rust
This command takes an already created canonical example variant and adds
an implementation to it. Additionally, if this is the first
implementation for the canonical example variant in question, it adds a
test implementation as well. This change only implements this
functionality for the /rust and /test directories - the follow up
implementation will do the same for all of the C++ directories
(/cpp_wire, /cpp_natural, and /hlcpp) as well.

Test: manual, used to generate fxr/737547
Bug: 111614
Change-Id: Ib136e71ed3f2eaf3fe7a92ebb0ec4999d9c46172
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737543
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>