[pbm] Fix errors on pb remove with no bundles
The attached bug reports on an error condition when removing
product-bundles if none are in the cache. This is just an uncaught error
mode when the list is empty, which is unhandled by the tool.

This CL explicitly handles this error condition and outputs it as
information rather than an error.

Bug: 111356
Change-Id: Ic2722548a3445077ba33ec7bff26135d723707b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736162
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>