[nxpfmac] Add support for new SDIO txn API
Now that the SDIO driver layer supports the new SDIO transaction API
switch the driver to use it for data frames. The new API is only used
for data transfers that allocate Frame objects from the components
library. This means it's currently only used for the data path. For
other transfers, such as firmware downloading, aggregated register
reads and so on, the old API is still used. Changing this requires
creating internal storage for data that is not allocated from
netdevice's frame pool.

Test: iperf3, fx test nxpfmac-sdio-tests, ACTS PingStressTest
Change-Id: I539439caa45103d1cf338d61c3820d7995d9c9e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736622
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>