[pbm] Prevent duplicate product bundle downloads
This CL adds a check to the pbms lib to prevent downloading a product
bundle that's already contained in the cache. In the case of invalid or
corrupt product bundles, we add a --force flag to the product bundle get
command to bypass the duplication check and download anyway.

There's also some error-reporting cleanup in the product-bundle tool.

Bug: 98281
Change-Id: Iaf58c6cc29e698a767f1ed93dd64bf4aeb51ccb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/735742
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>