[pbm] Clear cached metadata when removing bundles
The attached bug describes a use case where the user was confused by the
contents of the pbm cache after calling `ffx product-bundle remove`. The
root issue is that the metadata (product_bundles.json, etc.) is not
removed by the tool even if all of the bundles are, so they still show
up in a listing with --cached.

This CL adds logic to the remove command to clear the metadata for a
bundle if that bundle is the last one to be removed from a subdirectory.
In other words: since PBv1 bundles share a metadata file and directory,
we only remove those if all of the bundles it applies to are also
removed.

Bug: 111641
Change-Id: I341863a53a9c9f8145676ebbffb27e2895a08385
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737548
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>