[sdk] Fix sdk_cross_compile_host_tools implementation.
This configuration variable is used to decide whether to
cross-compile host tools when generating the SDK (on Linux).

It is currently defined to 'true' by default in //sdk/config.gni
but trying to set it to 'false' in args.gn results in no host
tools in the generated SDK, and a later failure when checking
its content from sdk/manifests/core.manifest.

This CL fixes the situation, meaning that when sdk_cross_compile_host_tool
is set to false in args.gn, then:

- The generated //sdk:core and //sdk:driver contents will not
  include arm64 host binaries on Linux. Note that these were
  never included on MacOS builds.

- The manifests named core.manifest and driver.manifest will
  be filtered to generate variants that ignore
  sdk://tools/<cpu>/ entries where <cpu> does not match the
  current host CPU architecture, and that will be used to
  verify the API.

  This feature requires a new boolean argument to the sdk() GN
  template named `filter_api_host_tools`

Bug: 111446
Change-Id: I4a12d783995abc9b247ff1087ab4ca1a7cdfa273
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736602
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>