[fidlc] Simplify deprecation checking
This CL simplifies the way fidlc checks that non-deprecated elements do
not refer to deprecated elements. Before, it used set operations for the
general case of arbitrary ranges. However, at that point in the code,
the elements are already decomposed so deprecation is guaranteed to be
all-or-nothing, making a much simpler check possible.

This also does some minor refactoring to get rid of
Availability::deprecated_range, which had only one other use.

Test: fx test fidlc-test
Change-Id: Idb9ad0440b624a13dc3a99e65e8bfdcb285f6c7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737205
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>