[netstack3] Drop `DeviceId::new_ethernet`
...to discourage tests from creating random `DeviceId`s and instead use
the `DeviceId` returned by creating a new device. This is because later
changes will back `DeviceId`s with a reference to the device instead of
a predictable scalar value used to lookup the device in a map.

Bug: 48578

Change-Id: I497b7b30c34b1f6d2c0ad5e6047db1e84ac1605b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737811
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alex Konradi <akonradi@google.com>