[emu] Add entropy to zbi from ffx emu
CL https://fxrev.dev/655762 added better entropy to the call to the ZBI
tool when launching an emulator from `fx emu`, but that change was never
conpied over to `ffx emu`. With the `fx emu` deprecation looming, this
CL catches the new launcher up, to make sure the change is not lost.

Change-Id: Ia782110686d37214ef7dccd8d75081001d02d116
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738404
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>