[fidlc][docs] Document ErrResponsesWithErrorsMustNotBeEmpty
fi-0076 and fi-0077 are a pair of error messages that explain the
different syntax required for response types when it should hold no
information. This is a bit inconsistent at the syntactic level but this
CL documents the current state.

Test: fx test fidlc-test; manual devsite staging
Bug: 111278
Change-Id: Icfa05d212f597ee147907c852e91d86362ea6821
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738006
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Matt Hamrick <mjhamrick@google.com>