[netstack] Add ctx to interface watcher event loop
Add a context.Context to the interface watcher event loop so that it
can be cancelled.

Add a unit test helper for starting an event loop and deferring a
function for waiting for the event loop goroutine to complete to avoid
leaking the goroutine and tripping the checker.

Test: fx test netstack-gotest

Change-Id: I9acb8a93785ff50a7e77e7a59dfb5da8dd965783
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734223
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>