[zxcrypt] Use DdkChildPreRelease to watch for removal
This is the officially-supported way to remove a child device and be
notified by device_manager when the device is about to be released.  We
can use this instead of rebinding the parent device, which will also
give us the ability to avoid replying to the `Seal()` call until
`device_manager` has also fully unbound the child, which means that a
subsequent `Unseal()` will not attempt to DdkAdd a child device with an
existing name.

Fixed: 111714
Change-Id: I476e607311e151765b8816073683885ff1aa6ac2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737988
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>