[docs] Document standard storage capability names.
This CL follows on from earlier conversations in the API and CF chat
rooms to define the standard names and behaviors we assign to the
commonly used storage capabilities.

These storage capabilities form a part of the API between a component
and the parent supplying its capabilities so Fuchsia.dev isn't the ideal
formal home for these definitions in the long term, but its better than
nothing.

Documenting these is a small but necessary step along the road to
agreeing a stash replacement so I'm tagging this with that bug.

Bug: 102879
Change-Id: Id0214806c1577f78ab155445bc0c218d26c38b48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734146
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>