[libc] Re-enable lsan test after fix
The test would sometimes fail because lsan could find the leaked pointer
on the stack. My hypothesis is that the leaked pointer was pushed onto
the stack in the initial allocation call somewhere in the "operator new"
lsan allocator call stack. On the next call to the runtime for checking
leaks, if the part of the stack where this pointer was pushed has not
been overwritten, then it will remain on the stack when the memory
snapshot scans the stack, which lsan will find and not treat as a leak.

16679fbda61fa11f1297df0db98cc384d37e5ab5 should address this flaky test
by scrubbing the stack after an allocation was made so the snapshot
cannot find it. I have not been able to reproduce the flake locally
again with this fix, so hopefully it won't appear on builders also.

Multiply: lsan-test: 100

Fixed: 111351
Change-Id: I6092f43d3bb9ef938524c9b44eafc8690ff99585
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738609
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Alex Brachet <abrachet@google.com>