[debugger] Print entry data in all contexts.
I noticed when doing some debugging that sym-debug commands failed to
print out DWARF expressions containing DW_OP_entry_value. This is
because there is no entry data provider which would normally provide the
registers for that use case.

But when printing the the DWARF program (not evaluating it), we do not
need the data and stringifying should always work. This change falls
back on the same data provider when printing entry values.

Change-Id: I9845bca36a37a731f6c937ffbb6543ffedf71b51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738652
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>