[zx][kazoo] Enforce proper FIDL naming in library zx
zx features inconsistent naming, much of which is invalid in the eyes of
the FIDL style guide. The library does not need to be a tall poppy in
this respect, and so this change
* updates the naming of almost every* declaration to be conventional, and
* enables the build-time enforcement of naming remaining conventional
going forward.

"Almost every" as there are a handful of snake case names exported in
the SDK in zx_common.fidl, some of which used prolifically (e.g.,
`zx.rights`) and more care will be needed to migrate these. For these
cases, there are already exemptions in the linter to ignore them.

Bug: 109734
Change-Id: Ie81aa619ca3d7372bf9fd896e7f67d66f445e0d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/725184
API-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>