[emu] Add deprecation warning and pause to `fx emu`
The `fx emu` tool will be retired at the end of October. This CL changes
the info banner about "early adoption" to a deprecation warning, and
increases the pause to a full 5 seconds. This gives the user more time
to observe the message, and encourages them to transition.

Bug: 98983
Change-Id: I6092ce0f6772d13b2dab7c0aec749236f8bdde57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738403
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>