[build][bazel] Fix update-workspace.py
Remove the hard-coded `out/default` that was used when the
build directory set with `fx set` was not found. This breaks
CI builders which do not use `fx set` and also use `out/not-default`
as the build directory.

+ Ensure that //build/bazel:generate_workspace always calls the
  script with the right build directory.

Bug: 111832
Change-Id: Ibc40a28b6b50abefcd45e0c8abe119ab9548d927
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739152
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>