[fdf] Pass drivers_serial_log test in DFv2
Pass //src/tests/drivers_serial_log:bringup_tests test with DFv2
enabled. The purpose of this test is to ensure the driver framework can
print logs to serial. In order to do this, the test waits for a certain
log statement to be printed that should be sent from the driver manager.
This log statement is not printed when DFv2 is enabled and so this test
must be updated to wait for a log statement from the driver manager
found in both DFv1 and DFv2.

Bug 106650

Change-Id: I192d736b6490962cf9c95935a966c13e75f1a0a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/736122
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>