[system-tests] Also retry IsDeviceUpToDate downgradeOTAAttempts times
The target may reboot during the "what version is on the target?"
pre-check as well as while we're trying to run the OTA, so try to be
resilient to that as well.  Using the same config flag makes sense here:
both are aiming to handle the downgrade build spontaneously rebooting.

Fixed: 110451
Bug: 109499
Change-Id: Iab679afaa851f09b3b48e713a0684c062700e310
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738445
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>