[build] Make json_validator targets host-only
This CL moves executable targets in //build/tools/json_validator into an
`if (is_host) { ... }` block to avoid expanding them in non-host
toolchains. Although these targets should not normally get built for
Fuchsia, somehow that was happening for some people, resulting in:

    ninja: error: dependency cycle: host_x64/json_validator -> host_x64/obj/sdk/lib/fidl_base/libfidl_base.a -> host_x64/obj/sdk/lib/fidl_base/libfidl_base.decoding_and_validating.cc.o -> obj/zircon/vdso/generate.stamp -> gen/include/zircon/syscalls/definitions.json -> fidling/obj/zircon/vdso/zx.stamp -> fidling/obj/zircon/vdso/zx_verify.stamp -> fidling/gen/zircon/vdso/zx_verify.json_validated -> host_x64/json_validator

The cycle occurs because it depends on fdio (because of configs applied
to all Fuchsia executables), which depends on FIDL bindings.

Change-Id: I63b3789461dfee3d399ddb7265f5b740f0de994e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739483
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>