[fidlc] Avoid duplicate errors for @available in unversioned library
This CL makes the AvailabilityStep mark an Avaibility as kFailed if it
bailed out because of ErrMissingLibraryAvailability. Previously it
didn't, so on visiting its children it would recursively try to visit
the parent again and hit the same error, making a duplicate message.

Test: fx test fidlc-test -- -f "VersioningTests.BadUseInUnversionedLibraryReportedOncePerAttribute"
Bug: 106549
Change-Id: Ib59af53fd2dcf7b139b32f13edc7c2578d163408
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738443
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>