[recovery] Add component that executes OTA updates.
Due to changes in isolated-ota and components v2 conversion, the OTA
functionality of recovery was broken. To fix this, the OTA capability
has been separated into a new component.

A separate component allows recovery to control the lifecycle of the OTA
process and ensure dependencies are created and managed appropriately.
This includes wiping blobfs and exposing its new directory at the right
time which would prove difficult in the main system_recovery component.

Test: fx test ota_unittests
Bug: b/243571575
Change-Id: Ia6bd3190c541036747a81b8c214f2dc0641cdc78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729343
Commit-Queue: Michael Brunson <mbrunson@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>