[netstack3] Remove rcv_residual for post FIN receipt states
This was added due to a concern for Core to be platform agnostic - For
example on a platform where netstack lives in the kernel, and we need to
keep all the bytes locally in stack, then we need such field to keep the
bytes that has been received but not read by the user yet for these
states. This is not a concern for Fuchsia at all and to support such a
platform doesn't seem to be near term. Removing it as a recent attempt
to remove the #[allow(unused)] lint reveals it is not being used. We can
add it back when the need arises.

Change-Id: I24de9e09c6ca2dc1b52054ef73f03ce284edf692
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739648
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>