[aml_mipicsi] Only join thread if created
thrd_t doesn't have a sentinal null value and if AmlMipiDevice::Create
fails, the thread is attempted to be joined with an invalid value
leading to a crash. Wrapping it in a std::optional allows us to know
whether or not it was initialized.

Change-Id: I7f650c861a6821eb789381bde13c402b954ee189
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738645
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>