[pbm] Reverse isatty conditions for output
The attached bug describes a bug in the product-bundle output, where the
`get` command is failing to print progress to the terminal. This is
because the check for a TTY output was reversed, so it would only print
on non-interactive outputs instead.

This CL switches the condition, so it will skip output only for
non-interactive sessions.

Bug: 111861
Change-Id: Ie950543d2ce35e650400f656be0ba20cd68bcee6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739563
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>