[pbm] Reordering PBM messages on duplicates
The attached bug points out that when attempting to `get` a duplicate
PBM, the tool was printing the duplicate repo message instead of the
duplicate bundle message. This was because the sequence of the checks
was reversed.

This CL puts an additional check for duplicate bundles before the check
for repos, so it's clear why the tool doesn't fetch the data. It will
still print the repo message if the user attempts to download a bundle
with the same name from a different version, as expected.

Bug: 111862
Change-Id: Ib4670091938d18af001907465420ca70925e54dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739266
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>