[pbms] RepoClient must update before refreshing
Originally RepoClient::new would auto-update the metadata to the latest
version before, but this was removed in a refactor. Unfortunately we
missed fixing this in the pbms library because there wasn't a test for
this functionality.

This patch moves the repository functionality into a dedicated module,
and adds the correct call to `RepoClient::update`. It also adds a test
to make sure an issue like this won't be missed in the future. We
unfortunately can't mock gcs, but this code was refactored to make it
mostly not dependent on the presence of gcs.

Fixed: 111508
Change-Id: Iec272dc011409ae003943d97060fbda0b7b3c0fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737854
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>