[pbm] Reverse sort order of product-bundle list
The attached bug points out how the output of `ffx product-bundle list`
sorts the bundles in alphanumeric order, which is the opposite of the
rest of the product-bundle commands and puts the newest bundles at the
bottom of the list.

This CL adds a call to reverse the order of the list, such that is
matches the rest of the product-bundle commands and puts the newest
bundles first.

Bug: 99226
Change-Id: I661f6e3ec2832d94407d91b4ae59d7e0643a5dbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739567
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>