[ffx] Remove ffx buildid from config.
This is a value that has no need of being looked up in the config. It is
currently used to determine whether or not the daemon is the same binary
as the one invoking it, and all of this is something that can be looked
up inside the binary itself.

Without this change, if ffx were to force a restart it would send its
runtime config, which gets really confusing if there are actually
different versions of ffx at play, as this buildid in the config would
not match the one of the actual binary.

Change-Id: I3c3f4e1e929ea4bfb026fa0d20d72732289bd75a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739654
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>