[pbm] List the bundles to be removed before-hand
The attached bug describes how it's not always clear which bundles are
about to be deleted by the remove command.

This CL adds code to list which bundles are passed to pb_remove_all()
when asking for confirmation, so the user has a chance to review before
indicating y/n.

Bug: 111769
Change-Id: I7cf9112a43ffdc14c19ebb0be8d20229bafbf107
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739366
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>