[pb] Add a transfer_manifest crate for upload/download of artifacts
This crate is reponsible for declaring the schema for the transfer
manifest, which is used to declare the local and remote paths for a set
of artifacts. This manifest will be used to upload all the artifacts for
a product bundle.

Bug: 111660
Change-Id: I1f79f5ee2551b83fe7590b910f787ef17b27c7e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/737862
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>