[settings] Add unit test to verify light controller calls without restore
In fxr/715324, an EnvironmentBuilder test was added to verify that Set
calls and the OnMicMute event don't panic when they're made after the
light controller times out and is started again. That test was a quick
test we added as the fix needed to go in quickly, so it indirectly tests
the behavior at the service-level and has to wait 6 seconds per test for
the controller to timeout.

This CL moves those tests over to a unit test that creates the
controller and calls methods directly, which no longer depends on the
controller discarding behavior.

The test itself is still pretty verbose and a good example of a test
that would benefit from mocking out objects.

Fixed: 107545
Bug: 103274
Test: setui_service_tests
Multiply: setui_service_tests
Change-Id: I722674906b9d19d4e9ece9e94e22762dd1134430
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738902
Reviewed-by: Sijie Chen <sijiec@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: William Xiao <wxyz@google.com>