[netstack3] Remove #[allow(unused)] attribute on TCP mod
This attribute was added to suppress errors at an early stage when not
much from the TCP module was referenced by Bindings. Now that we are
actually using it, we can remove this and fix mistakes that have slipped
through while this attribute was on.

Fixed: 95688
Change-Id: I42b49d3df62d0ab34d5fe6aa74ecb3759171a023
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739443
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>