[credential manager] Log error for the synchonization failure.
With the problem in fxb/111926 I was running into this. Prior to this CL
if the disk and Cr50 get out of sync credential manager just silently
dies without logging anything. This doesn't fix any of the root causes
but at least if we're looking at logs where people were in a
missynchronized state we'll know it.

Change-Id: I5576e8ed2c145dee86997c0a01dbcde305710a79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739780
Fuchsia-Auto-Submit: Jody Sankey <jsankey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ben Wright <benwright@google.com>