[usb-cdc-ecm] Use C++ Request data structures
Also remove the use of asynchronous endpoint resets, which complicate
the code for little benefit. The driver was previously holding a lock
across a usb request lifecycle to keep the resets from interleaving
anyway, so doing it synchronously is not a significant behaviour change.

Change-Id: I73425bf4a16adb410c9f169b05d66b6bdee53a98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723727
Commit-Queue: Alex Legg <alexlegg@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>