[docgen] Write function default values.
Clang-doc now emits the default values for function parameters if set in
the original code. This is important to know how to call the function.

Default value information is now emitted in the function declarations
generated by cppdocgen.

Avoid duplicating the title when a custom title is supplied for a
function or defined.

Rename "std::basic_string" to "std::string" when printing types.

Remove extra space in enum definitions.

Change-Id: I8295c1dd93819b7ae1c4fab15af943a812e06eb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738723
Reviewed-by: Dave Smith <smithdave@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>