[ffx][component] Implement copy from target file size limit error
Currently, if a user tries to copy a file from target that has a file
size greater than the zircon channel limit of 64KB, 64KB of data will be
copied over to the host. This can result in unexpected behaviour. To fix
this issue, this CL retrieves the file size from target and throws an
error if it is above 64KB. Additionally, this CL fixes an issue with an
incorrect usage of the test_case module and now properly verifies the
copy_from_device_to_host tests.

Future CLs will implement copy support for file sizes greater than 64KB.

Bug: 100465
Bug: 111473

Change-Id: If163e145f7cb1f37d769ba976d2f2fdd37376e51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739573
Commit-Queue: Richard Chen <richardrchen@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>