[brcmfmac] Make SDIO func1 lock more granular
The SDIO func1 lock has become a lock protecting more than just func1
access at this point. The lock was held for the entire IRQ DPC,
including the part that keeps looping if something triggers the DPC
while the DPC is running. In cases where we run out of receive buffers
the DPC would remain triggered causing this loop to spin endlessly
without releasing the lock. This caused other threads, notable the main
work thread, to deadlock because it was trying to acquire the lock to
retrieve stats and other things.

Improve the situation by acquiring the lock inside the loop for each DPC
call instead. This seems to resolve the deadlock, allowing the main work
thread to continue working.

Test: Manual stress testing
Bug: 111866
Change-Id: Id06b2231a85b1f48815c0dcfc2291dcdc811b033
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740578
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>