[docgen] Format links in comments.
Looks for markdown links with no destination and attempts to look up the
symbol name and automatically create the proper link destination for it.

Move the existing comment formatting code to a new file since it's
getting more complexity.

Enhance the documentation. Remove the obsolete instructions on running
separately and document how docstrings and titles work, along with the
new link behavior.

Make the _toc writing code order stable. Previously this was in map
order which is based on hash functions.

The current code does not handle non-toplevel symbols (e.g. with classes
and namespaces, or anything with a "::" in it). This will be addressed
in a followup.

Change-Id: I68e929599e1fe10a051207340d5084cae4e15979
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739444
Reviewed-by: Dave Smith <smithdave@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>