[kernel][x86] make sure top level page table is <4GB when bootstrapping secondary cpus
In specific situations (KASAN on machines with low <4GB memory) by the
time the bootstrap aspace is created all of low memory may be exhausted.
This causes problems when bootstrapping secondary cpus as they go
through the real mode -> 32bit protected mode -> 64bit protected mode
bootstrap.

Make a copy of the top level page table for the bootstrap aspace in low
memory just prior to starting the secondary cores to make sure this
constraint is met. Since the aspace is static at this point and this
copy of the top level page is only used as the cpus bounce through the
modes it is safe to use a copy.

Bug: 96565
Change-Id: I36776585c3812654a187c11f88d898c3370b744f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739884
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>