[check-licenses] Build API contract updates.
Move the build_api_module target out of the check-licenses directory.
Other projects define theirs in the root //BUILD.gn file, so I will do
the same in a future CL.

Also:
Previously, license files and segments were stored in a subdirectory of
the "license" folder, in the output directory. By default it still is,
but when uploading these files to GCS, it's probably better to remove
paths / details that are implementation-specific to check-licenses.

This CL adds an optional path argument, when provided it will save the
license files to that path instead of the default one.

Change-Id: I291da6ed984fcb182cb48d55deac1b9f1b40fdf0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740369
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>